requirement.txt
numpy
flask
matplotlib
PILLOW
requests
scikit-learn
torch
tqdm

